requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
configparser==6.0.0
curl_cffi==0.6.0b9
distro==1.9.0
dnspython==2.6.0
duckduckgo_search==4.4.3
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
lxml==5.1.0
nest-asyncio==1.6.0
openai==1.12.0
outcome==1.3.0.post0
packaging==23.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
PySocks==1.7.1
python-docx==1.1.0
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.1
wsproto==1.2.0
